ultimos cambios

+++ src/main/java/com/users/users/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.users.users.controllers;
 
 import com.users.users.dao.UserDao;
 import com.users.users.models.User;
+import com.users.users.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,9 +14,16 @@ public class UserController {
 
     @Autowired
     private UserDao userDao;
+    @Autowired
+    private JWTUtil jwtUtil;
 
+    private boolean validarToken(String token) {
+        String usuarioId = jwtUtil.getKey(token);
+        return usuarioId != null;
+    }
     @RequestMapping(value = "api/users", method = RequestMethod.GET)
-    public List<User> getUsers(){
+    public List<User> getUsers(@RequestHeader(value="Authorization") String token){
+        if (!validarToken(token)) { return null; }
        return userDao.getUsers();
     }
     @RequestMapping(value = "api/users", method = RequestMethod.POST)
@@ -27,7 +35,8 @@ public class UserController {
         userDao.register(user);
     }
     @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable Long id){
+    public void delete(@RequestHeader(value="Authorization") String token, @PathVariable Long id){
+        if (!validarToken(token)) { return; }
         userDao.delete(id);
     }
